Доработана система заполнения пинборда. Создана ссылка "Мой Профиль". Доработана система администраторов с доступом к редактированию данных пользователя и его пинборда.

+++ Controllers/ProfileController.cs
@@ -1,6 +1,8 @@
 ﻿using System.Net;
+using Azure.Core;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Primitives;
 using QualificationSite.Services.Interfaces;
 using QualificationSite.ViewModel;
 
@@ -21,6 +23,13 @@ public class ProfileController : Controller
         this.service = service;
         this.appEnvironment = appEnvironment;
     }
+
+    [HttpGet]
+    public async Task<IActionResult> My()
+    {
+        var id = await db.GetIdByNameAsync(HttpContext.User.Identity.Name);
+        return Redirect($"/Profile/{id}");
+    }
     [HttpGet]
     public async ValueTask<IActionResult> Id(long id) // ValueTask на тот случай, если не придётся использовать асинхронный код
     {
@@ -28,25 +37,26 @@ public class ProfileController : Controller
             return RedirectToAction("Login", "Account");
         ViewBag.Path = appEnvironment.WebRootPath;
         var pins = await service.GetPinsAsync(id);
+        var profile = await service.GetProfileAsync(id);
+        ProfileViewModel model;
         if (HttpContext.User.IsInRole("Admin"))
         {
-            var response = await service.GetProfileAsync(id);
+            var response = profile;
             if (response.StatusCode == HttpStatusCode.Found)
             {
                 ViewBag.EditAccess = true;
-                ProfileViewModel model = new ProfileViewModel
+                model = new ProfileViewModel
                 {
                     Age = response.Data.Age.Value, City = response.Data.City, Languages = response.Data.Languages,
                     Name = response.Data.Name, Surname = response.Data.Surname, University = response.Data.University,
-                    Pins = pins.Data, Id = response.Data.Id
+                    Pins =  pins.Data, Id = response.Data.Id
                 };
                 return View(model);
             }
         }
         if (await db.GetNameByIdAsync(id) == HttpContext.User.Identity.Name)
         {
-            var response = await service.GetProfileAsync(id);
-            ProfileViewModel model;
+            var response = profile;
             if (response.StatusCode == HttpStatusCode.Found)
             {
                 ViewBag.EditAccess = true;
@@ -58,6 +68,11 @@ public class ProfileController : Controller
                 };
                 return View(model);
             }
+            return RedirectToAction("CreateProfile");
+        }
+        if (profile.StatusCode == HttpStatusCode.Found)
+        {
+            var response = profile;
             model = new ProfileViewModel
             {
                 Age = response.Data.Age.Value, City = response.Data.City, Languages = response.Data.Languages,
@@ -66,12 +81,13 @@ public class ProfileController : Controller
             };
             return View(model);
         }
-        return RedirectToAction("AccessRejected");
+        else
+            return RedirectToAction("AccessRejected");
     }
     [HttpGet]
     public async Task<IActionResult> Edit(long id)
     {
-        if (await db.GetIdByNameAsync(HttpContext.User.Identity.Name) != id)
+        if (await db.GetIdByNameAsync(HttpContext.User.Identity.Name) != id && !HttpContext.User.IsInRole("Admin"))
             return RedirectToAction("AccessRejected");
         var response = await service.GetProfileAsync(id);
         if (response.StatusCode == HttpStatusCode.Found)
@@ -82,10 +98,15 @@ public class ProfileController : Controller
     [HttpPost]
     public async Task<IActionResult> Edit(ProfileViewModel model)
     {
+        long profileId;
+        string str = HttpContext.Request.Form["profileId"].First();
+        profileId = long.Parse(str.Replace("/", ""));
         var id = await db.GetIdByNameAsync(HttpContext.User.Identity.Name);
+        if (id != profileId && !HttpContext.User.IsInRole("Admin"))
+            return RedirectToAction("AccessRejected");
         if (ModelState.IsValid)
-            await service.EditProfile(model, id); // TODO: сделать обработку неверной модели.
-        return Redirect("/profile/" + id);
+            await service.EditProfile(model, profileId); // TODO: сделать обработку неверной модели.
+        return Redirect("/profile/" + profileId);
     }
     public IActionResult AccessRejected() => View();
     [HttpGet]
@@ -110,7 +131,7 @@ public class ProfileController : Controller
     {
         if (ModelState.IsValid)
             await service.EditProfile(model, await db.GetIdByNameAsync(HttpContext.User.Identity.Name));
-        return RedirectToAction("Index", "Home");
+        return await My();
     }
 
     [HttpGet]
@@ -129,16 +150,25 @@ public class ProfileController : Controller
         var response = await service.GetPinAsync(id);
         if (response.StatusCode == HttpStatusCode.NotFound)
             return RedirectToAction("AccessRejected");
-        await service.DeletePinAsync(response.Data);
-        return RedirectToAction("Index", "Home");
+        var profileId = await db.GetIdByNameAsync(HttpContext.User.Identity.Name);
+        if ((await service.IsUserPinAsync(response.Data, profileId)).Data || HttpContext.User.IsInRole("Admin"))
+            await service.DeletePinAsync(response.Data);
+        else
+            return RedirectToAction("AccessRejected");
+        return Redirect("/profile/editpinboard/" + profileId);
     }
 
     [HttpPost]
     public async Task<IActionResult> CreatePin(ProfilePinViewModel model)
     {
+        long profileId;
+        string str = HttpContext.Request.Form["profileId"].First();
+        profileId = long.Parse(str.Replace("/", ""));
         var id = await db.GetIdByNameAsync(HttpContext.User.Identity.Name);
-        model.ProfileId = id;
+        if (id != profileId && !HttpContext.User.IsInRole("Admin"))
+            return Redirect("/profile/editpinboard/" + profileId); 
+        model.ProfileId = profileId;
         await service.CreatePinAsync(model);
-        return RedirectToAction("Index", "Home");
+        return Redirect("/profile/editpinboard/" + profileId);
     }
 }
